Change test package name
