+++ src-test/org/graphstream/ui/viewer/test/DemoTwoGraphsInOneViewerFx.java
@@ -29,7 +29,7 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C and LGPL licenses and that you accept their terms.
  */
-package org.graphstream.ui.viewer_fx.test;
+package org.graphstream.ui.viewer.test;
 
 import java.io.IOException;
