+++ src-test/org/graphstream/ui/viewer/test/data/dorogovtsev_mendes6000.dgs
